+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- (CYBRSOURCE-33) Convert Chile region name to iso code
+
 ## [1.3.0] - 2022-06-15
 
 ### Fixed